+++ packages/gateway/src/metric/pubsub-events.enum.ts
@@ -1,7 +1,0 @@
-enum PubSubEventsEnum {
-  RESOLVER_CALLED = 'resolverCalled',
-  RESOLVER_DONE = 'resolverDone',
-  RESOLVER_ERROR = 'resolverError',
-}
-
-export default PubSubEventsEnum;