+++ packages/gateway/src/middleware/log-response-error.ts
@@ -1,5 +1,6 @@
 import { ErrorRequestHandler } from 'express';
-import { metricEmitter, MetricsChannels } from '../metric';
+import { MetricsChannels } from '@graphql-portal/types';
+import { metricEmitter } from '../metric';
 
 export const logResponseError: ErrorRequestHandler = (err, req, res, next) => {
   if (err) metricEmitter.emit(MetricsChannels.GOT_ERROR, req.id, err);