+++ packages/config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphql-portal/config
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [d61e610]
+- Updated dependencies [1c86dc9]
+  - @graphql-portal/types@0.7.0
+
 ## 0.5.0
 
 ### Minor Changes